+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Replaced the default GMP-style primality check with the BPSW'21 check. ([#70])
 - Removed `is_prime_with_rng()` and `is_safe_prime_with_rng()` since the primality check is now deterministic. ([#70])
+- `SieveIterator` removed from the public API. ([#79])
 
 
 ### Added
@@ -21,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [#70]: https://github.com/entropyxyz/crypto-primes/pull/72
 [#72]: https://github.com/entropyxyz/crypto-primes/pull/72
+[#79]: https://github.com/entropyxyz/crypto-primes/pull/79
 
 
 ## [0.7.0-pre.0] - 2025-02-22